# support compactifications of more sophisticated types