--- a/src/affectedqubits.jl
+++ b/src/affectedqubits.jl
@@ -12,9 +12,9 @@ affectedqubits(p::PauliOperator) = 1:length(p)
 affectedqubits(m::Union{AbstractMeasurement,sMRX,sMRY,sMRZ}) = (m.qubit,)
 affectedqubits(v::VerifyOp) = v.indices
 affectedqubits(c::CliffordOperator) = 1:nqubits(c)
-affectedqubits(c::ClassicalXORConcreteWorkaround) = ()
+affectedqubits(c::ClassicalXOR) = ()
 
 affectedbits(o) = ()
 affectedbits(m::sMRZ) = (m.bit,)
 affectedbits(m::sMZ) = (m.bit,)
-affectedbits(c::ClassicalXORConcreteWorkaround) = (c.bits..., c.store)
+affectedbits(c::ClassicalXOR) = (c.bits..., c.store)